services/requirements.txt
pip==22.2.2
setuptools==63.2.0
bottle
black==22.3.0
requests
python-dotenv
sqlalchemy
mysqlclient
bottle-websocket
websocket-client
websockets
gevent
gevent-websocket